LayoutTests/webanimations/animation-current-time-expected.txt

Check that the Animation currentTime property is set based on the timeline currentTime and the animation startTime.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Create an animation that starts at 2s.
PASS animation.currentTime is null

Seek the timeline to 0s.
PASS animation.currentTime is -2

Seek the timeline to 3s.
PASS animation.currentTime is 1

Set the animation start time to 1s.
PASS animation.currentTime is 2

Setting the current time to null should throw a TypeError.
PASS animation.currentTime = null threw exception TypeError: Type error.

Setting the current time should adjust the start time.
PASS animation.startTime is 2

PASS successfullyParsed is true

TEST COMPLETE
